Claude-Dooku/pyproject.toml
[project]
name = "claude-dooku"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.10.1",
    "python-dotenv>=1.2.1",
]
